% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{as_data_uri}
\alias{as_data_uri}
\title{File or any R objects to Data URI, base64 encoded}
\usage{
as_data_uri(obj, mime = "text/plain;charset=US-ASCII", is_file = FALSE,
  ...)
}
\description{
File or any R objects to Data URI, base64 encoded
}
