% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{as_data_uri_file}
\alias{as_data_uri_file}
\title{Convert file to dataURI via base64enc package}
\usage{
as_data_uri_file(file, ...)
}
\arguments{
\item{file}{file to be converted}

\item{...}{pass to base64enc::dataURI}
}
\description{
Convert file to dataURI via base64enc package
}
