% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{geom_plane}
\alias{geom_plane}
\title{Create plane object (GeomPlane)}
\usage{
geom_plane(position, mesh_name, width, height, mesh_info = mesh_name,
  layer = 1, hover_enabled = T, is_clipper = F, ...)
}
\arguments{
\item{position}{vector of 3, geom coordinates, for example c(0,0,0) stands for origin}

\item{mesh_name}{character, unique identifier/name,}

\item{width}{plane width}

\item{height}{plane height}

\item{mesh_info}{default is mesh_name, html info to show when object is clicked}

\item{layer}{geom layer (visibility), see ?threejs_scene}

\item{hover_enabled}{when mouse hover over, show helpers}

\item{is_clipper}{can clip objects?}

\item{...}{other params for GeomPlane$new(...)}
}
\description{
Create plane object (GeomPlane)
}
