% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{geom_sphere}
\alias{geom_sphere}
\title{Create sphere object (GeomSphere)}
\usage{
geom_sphere(position, mesh_name, radius, mesh_info = mesh_name,
  layer = 1, hover_enabled = T, ...)
}
\arguments{
\item{position}{vector of 3, geom coordinates, for example c(0,0,0) stands for origin}

\item{mesh_name}{character, unique identifier/name,}

\item{radius}{number, positive}

\item{mesh_info}{default is mesh_name, html info to show when object is clicked}

\item{layer}{geom layer (visibility), see ?threejs_scene}

\item{hover_enabled}{when mouse hover over, show helpers}

\item{...}{other params for GeomSphere$new(...)}
}
\description{
Create sphere object (GeomSphere)
}
