% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_freemesh.R
\docType{data}
\name{GeomFreeMesh}
\alias{GeomFreeMesh}
\title{Free Geom object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeomFreeMesh
}
\description{
Free Geom object
}
\keyword{datasets}
