% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.freesurf.R
\name{read.freesurf.asc}
\alias{read.freesurf.asc}
\title{Read free-surfer files (ascii)}
\usage{
read.freesurf.asc(file)
}
\arguments{
\item{file}{file to be read}
}
\description{
Read free-surfer files (ascii)
}
