% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.freesurf.R
\name{read.freesurf.gii}
\alias{read.freesurf.gii}
\title{Read free-surfer files (gifti)}
\usage{
read.freesurf.gii(file)
}
\arguments{
\item{file}{file to be read}
}
\description{
Read free-surfer files (gifti)
}
