% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_scene.R
\name{renderThreejs}
\alias{renderThreejs}
\title{Render shiny-threejsr scene}
\usage{
renderThreejs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{env}{runtime environment see shiny outputs}

\item{quoted}{is expr quoted?}
}
\description{
Render shiny-threejsr scene
}
