% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_scene.R
\name{threejs_scene}
\alias{threejs_scene}
\title{Main function to render threejsr scene}
\usage{
threejs_scene(...)
}
\arguments{
\item{...}{pass to other methods}
}
\description{
Main function to render threejsr scene
}
