% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_scene.R
\name{threejsOutput}
\alias{threejsOutput}
\title{Shiny-threejsr output}
\usage{
threejsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{character ID for shiny outputs}

\item{width}{css width default '100\%'}

\item{height}{css height default '400px}
}
\description{
Shiny-threejsr output
}
