% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{threejsr_palettes}
\alias{threejsr_palettes}
\title{Util function to map data to color}
\usage{
threejsr_palettes(name, nlevels = 64, plot = F, customize, data,
  center, bias = 1, space = "rgb", interpolate = "linear",
  reversed = FALSE)
}
\arguments{
\item{name}{color scheme name, choices are "rainbow", "grey", "terrain", "heat", "cm", or "customized"}

\item{nlevels}{number of distinct colors}

\item{plot}{visualize color mappings}

\item{customize}{vector of colors, see grDevices::colorRampPalette (valid only when name="customized")}

\item{data}{any vector, matrix or array to calculate color mappings (can be blank)}

\item{center}{center alignmnet. For example, if you want color center standing for data value 0, set center=0. Default is the mean of range(data)}

\item{bias}{passed to grDevices::colorRampPalette, see grDevices::colorRampPalette}

\item{space}{passed to grDevices::colorRampPalette, see grDevices::colorRampPalette}

\item{interpolate}{passed to grDevices::colorRampPalette, see grDevices::colorRampPalette}

\item{reversed}{reverse color map}
}
\description{
Util function to map data to color
}
